refactor: 상수와 유틸을 helpers에서 lib로 수정
